This commit fixes issue #7, in which there was an error when loading social images from the page: https://piscinadeentropia.es/about/about_contacto/.
The error was caused by a bad link to the images, this only occurred here because being HTML code MkDocs did not rewrite the paths of the images correctly, however the photo below continued to be displayed because the link was in Marckdown format.

For example, this problem does not occur on the home page (index.md), because the page is located in the main directory, above everything else. However, the contact page is located inside a directory and when Mkdocs converts the files so that the web server can serve them it moves them into another directory with the name of the page.

The solution I have decided to implement, is to access the images directly from the URL of the web and not from an internal link like for example:

Old link(Internal): ../images/pisicina_de_entropia_logo.png
New (External) link:https://piscinadeentropia.es/images/pisicina_de_entropia_logo.png