Update mkdocs.yml
This commit fix the #9 Issue, that found a versioning error in the MkDocs logs.

This warning was generated by MkDocs because MkDocs.yml did not specify which version of the theme to use, it was not an important error, however it is interesting to fix it. The bug can be easily fixed by adding `version:` to the theme section of mkdocs.yml.

Signed-off-by: Isaac Hernán <77550577+Isaaker@users.noreply.github.com>